feat(search): signed --max-count -- negative N keeps the latest N
--max-count is now signed: N > 0 keeps the EARLIEST N exchanges of the
chronological stream (unchanged), -N keeps the LATEST N, and 0 stays uncapped
(the crate-wide convention). The kept exchanges always emit oldest-first among
themselves -- one ordering rule; the sign only selects which end of the sorted
timeline survives, mirroring the shared range grammar's -k = from-the-end
form (--turn -3..).

This makes "the most recent occurrence of X" a one-liner
(csift search "X" --max-count -1) alongside the existing earliest form
(--max-count 1).

Both ends disclose the window: the head banner appends
"showing earliest/latest K", and the footer's drop report now names the
dropped side ("235 earlier dropped by --max-count" under a latest-N window,
"later" under an earliest-N one; the --raw stderr note matches). show/list/
stats --max-count are unchanged -- they address records, not a chronological
merge.

+++ src/cli.rs
@@ -1177,10 +1177,14 @@ pub struct SearchArgs {
     #[arg(long, value_name = "WHEN")]
     pub until: Option<String>,
 
-    /// Cap emitted exchanges (default: unlimited — no cap). NO silent truncation — the drop
-    /// count is reported.
-    #[arg(long, value_name = "N")]
-    pub max_count: Option<usize>,
+    /// Cap emitted exchanges — SIGNED: `N` keeps the EARLIEST N of the chronological stream,
+    /// `-N` keeps the LATEST N (so `--max-count 1` = the first occurrence ever, `--max-count -1`
+    /// = the most recent one), `0` = uncapped (default: unlimited). The kept exchanges always
+    /// emit oldest-first among themselves — the sign only selects which END of the stream
+    /// survives, mirroring the shared range grammar's `-k = from the end`. NO silent
+    /// truncation — the head banner shows the emitted window, the footer reports the drop.
+    #[arg(long, value_name = "N", allow_negative_numbers = true)]
+    pub max_count: Option<i64>,
 
     /// Print ONLY the total number of matching exchanges (one integer) — the ripgrep
     /// `-c` idiom for "how many times X?". Honors every filter (`-t`, time window,
